fix delete chat

+++ chat/internal/repository/chats/repository.go
@@ -40,7 +40,7 @@ func (p *PostgresChatRepository) CreateChat(ctx context.Context, chatInfo *model
 	return converter.MsgRepoToService(chatRep), nil
 }
 
-func (p *PostgresChatRepository) DeleteChat(ctx context.Context, idChat int) error {
+func (p *PostgresChatRepository) DeleteChat(ctx context.Context, idChat int64) error {
 	builderDelete := sq.Delete("chats").PlaceholderFormat(sq.Dollar).Where(sq.Eq{"id_chat": idChat})
 
 	query, args, err := builderDelete.ToSql()
