+++ chat/internal/repository/repository.go
@@ -14,5 +14,5 @@ type MessageRepository interface{
 
 type ChatRepository interface{
 	CreateChat(ctx context.Context, chatInfo *model.ChatService) (*model.ChatService, error)
-	DeleteChat(ctx context.Context, idChat int) error
+	DeleteChat(ctx context.Context, idChat int64) error
 }
