+++ chat/internal/service/chat/deleteChat.go
@@ -4,6 +4,6 @@ import(
 	"context"
 )
 
-func (s *servChat) DeleteChat(ctx context.Context, idChat int) error{
+func (s *servChat) DeleteChat(ctx context.Context, idChat int64) error{
 	return s.chatRepo.DeleteChat(ctx, idChat)
 }
